Refactor utils
Renamed src/headless/utils/core.js -> src/headless/utils/index.js
Move functions that rely on the `_converse` object to src/headles/utils/session.js
Move object-related functions to src/headless/utils/object.js
Remove `contains` and `contains.not` functions, they're not being used.
