+++ dev.html
@@ -40,8 +40,8 @@
         muc_show_logs_before_join: true,
         notify_all_room_messages: ['discuss@conference.conversejs.org'],
         view_mode: 'fullscreen',
-        // websocket_url: 'wss://conversejs.org/xmpp-websocket',
-        websocket_url: 'ws://chat.example.org:5380/xmpp-websocket',
+        websocket_url: 'wss://conversejs.org/xmpp-websocket',
+        // websocket_url: 'ws://chat.example.org:5380/xmpp-websocket',
         whitelisted_plugins: ['converse-debug'],
         // connection_options: { worker: '/dist/shared-connection-worker.js' }
     });
