+++ src/headless/plugins/bosh.js
@@ -10,6 +10,8 @@ import log from "../log.js";
 import { BOSH_WAIT } from '../shared/constants.js';
 import { Model } from '@converse/skeletor/src/model.js';
 import { setUserJID, } from '../utils/init.js';
+import { isTestEnv } from '../utils/session.js';
+import { createStore } from '../utils/storage.js';
 
 const { Strophe } = converse.env;
 
@@ -33,7 +35,7 @@ converse.plugins.add('converse-bosh', {
             const id = BOSH_SESSION_ID;
             if (!_converse.bosh_session) {
                 _converse.bosh_session = new Model({id});
-                _converse.bosh_session.browserStorage = _converse.createStore(id, "session");
+                _converse.bosh_session.browserStorage = createStore(id, "session");
                 await new Promise(resolve => _converse.bosh_session.fetch({'success': resolve, 'error': resolve}));
             }
             if (_converse.jid) {
@@ -93,7 +95,7 @@ converse.plugins.add('converse-bosh', {
                     _converse.connection.restore(jid, _converse.connection.onConnectStatusChanged);
                     return true;
                 } catch (e) {
-                    !_converse.isTestEnv() && log.warn("Could not restore session for jid: "+jid+" Error message: "+e.message);
+                    !isTestEnv() && log.warn("Could not restore session for jid: "+jid+" Error message: "+e.message);
                     return false;
                 }
             }
