+++ src/headless/plugins/disco/entity.js
@@ -5,6 +5,7 @@ import sizzle from 'sizzle';
 import { Collection } from '@converse/skeletor/src/collection';
 import { Model } from '@converse/skeletor/src/model.js';
 import { getOpenPromise } from '@converse/openpromise';
+import { createStore } from '../../utils/storage.js';
 
 const { Strophe } = converse.env;
 
@@ -28,21 +29,21 @@ class DiscoEntity extends Model {
 
         this.dataforms = new Collection();
         let id = `converse.dataforms-${this.get('jid')}`;
-        this.dataforms.browserStorage = _converse.createStore(id, 'session');
+        this.dataforms.browserStorage = createStore(id, 'session');
 
         this.features = new Collection();
         id = `converse.features-${this.get('jid')}`;
-        this.features.browserStorage = _converse.createStore(id, 'session');
+        this.features.browserStorage = createStore(id, 'session');
         this.listenTo(this.features, 'add', this.onFeatureAdded);
 
         this.fields = new Collection();
         id = `converse.fields-${this.get('jid')}`;
-        this.fields.browserStorage = _converse.createStore(id, 'session');
+        this.fields.browserStorage = createStore(id, 'session');
         this.listenTo(this.fields, 'add', this.onFieldAdded);
 
         this.identities = new Collection();
         id = `converse.identities-${this.get('jid')}`;
-        this.identities.browserStorage = _converse.createStore(id, 'session');
+        this.identities.browserStorage = createStore(id, 'session');
         this.fetchFeatures(options);
     }
