+++ src/headless/plugins/mam/placeholder.js
@@ -1,5 +1,5 @@
 import { Model } from '@converse/skeletor/src/model.js';
-import { getUniqueId } from '../../utils/core.js';
+import { getUniqueId } from '../../utils/index.js';
 
 export default class MAMPlaceholderMessage extends Model {
