+++ src/headless/plugins/muc/utils.js
@@ -2,7 +2,7 @@ import _converse from '../../shared/_converse.js';
 import api, { converse } from '../../shared/api/index.js';
 import log from '../../log.js';
 import { ROLES } from './constants.js';
-import { safeSave } from '../../utils/core.js';
+import { safeSave } from '../../utils/index.js';
 
 const { Strophe, sizzle, u } = converse.env;
