+++ src/plugins/chatview/message-form.js
@@ -3,7 +3,7 @@ import { ElementView } from '@converse/skeletor/src/element.js';
 import { __ } from 'i18n';
 import { _converse, api, converse } from "@converse/headless";
 import { parseMessageForCommands } from './utils.js';
-import { prefixMentions } from '@converse/headless/utils/core.js';
+import { prefixMentions } from '@converse/headless/utils/index.js';
 
 const { u } = converse.env;
