+++ src/plugins/fullscreen/index.js
@@ -4,7 +4,7 @@
  * @copyright 2022, the Converse.js contributors
  */
 import { api, converse } from "@converse/headless";
-import { isUniView } from '@converse/headless/utils/core.js';
+import { isUniView } from '@converse/headless/utils/session.js';
 
 import './styles/fullscreen.scss';
