+++ src/plugins/minimize/utils.js
@@ -1,5 +1,6 @@
 import { _converse, api, converse } from '@converse/headless';
 import { __ } from 'i18n';
+import { isTestEnv } from '@converse/headless/utils/session.js';
 
 const { dayjs, u } = converse.env;
 
@@ -61,7 +62,7 @@ function getBoxesWidth (newchat) {
  * @param { _converse.ChatBoxView|_converse.ChatRoomView|_converse.ControlBoxView|_converse.HeadlinesFeedView } [newchat]
  */
 export function trimChats (newchat) {
-    if (_converse.isTestEnv() || api.settings.get('no_trimming') || api.settings.get("view_mode") !== 'overlayed') {
+    if (isTestEnv() || api.settings.get('no_trimming') || api.settings.get("view_mode") !== 'overlayed') {
         return;
     }
     const shown_chats = getShownChats();
