+++ src/plugins/rosterview/templates/group.js
@@ -2,7 +2,7 @@ import 'shared/components/icons.js';
 import { __ } from 'i18n';
 import { _converse, converse } from "@converse/headless";
 import { html } from "lit";
-import { isUniView } from '@converse/headless/utils/core.js';
+import { isUniView } from '@converse/headless/utils/session.js';
 import { repeat } from 'lit/directives/repeat.js';
 import { toggleGroup } from '../utils.js';
